FIles refactored, provided get removed in limit solvers
