+++ Sources/AngouriMath/Functions/Continuous/Limits/Solvers/Limit.Classes.cs
