+++ Sources/AngouriMath/Functions/Continuous/Limits/Transformations.cs
@@ -1,0 +1,94 @@
+﻿/* 
+ * Copyright (c) 2019-2021 Angouri.
+ * AngouriMath is licensed under MIT. 
+ * Details: https://github.com/asc-community/AngouriMath/blob/master/LICENSE.md.
+ * Website: https://am.angouri.org.
+ */
+using System;
+using static AngouriMath.Entity;
+
+namespace AngouriMath.Functions.Algebra
+{
+    partial class LimitFunctional
+    {
+        private static Entity EquivalenceRules(Entity expr, Variable x, Entity dest)
+            => expr switch
+            {
+                Sinf or Tanf or Arcsinf or Arctanf => expr.DirectChildren[0],
+                _ => expr
+            };
+
+        private static Entity ApplyFirstRemarkable(Entity expr, Variable x, Entity dest)
+            => expr switch
+            {
+                Divf(var a, var b) div
+                    when a.Limit(x, dest).Evaled == 0 && b.Limit(x, dest).Evaled == 0
+                        => div.New(EquivalenceRules(a, x, dest), EquivalenceRules(b, x, dest)),
+
+                _ => expr
+            };
+
+        private static Entity ApplySecondRemarkable(Entity expr, Variable x, Entity dest)
+            => expr switch
+            {
+                // f(x)^g(x) for f(x) -> 1, g(x) -> +oo
+                // => (1 + (f(x) - 1)) ^ g(x) = ((1 - (f(x) - 1)) ^ (1 / (f(x) - 1))) ^ (g(x) (f(x) - 1))
+                // e ^ (g(x) * (f(x) - 1))
+                Powf(var xPlusOne, var xPower) when
+                xPlusOne.ContainsNode(x) && xPower.ContainsNode(x) &&
+                (xPlusOne - 1).Limit(x, dest).Evaled == 0 && IsInfiniteNode(xPower.Limit(x, dest)) =>
+                MathS.e.Pow(xPower * (xPlusOne - 1)),
+
+                _ => expr
+            };
+
+        private static bool IsInfiniteNode(Entity expr)
+            => expr.ContainsNode("+oo") || expr.ContainsNode("-oo"); // TODO: is it correct?
+
+        private static bool IsFiniteNode(Entity expr)
+            => !IsInfiniteNode(expr) && expr != MathS.NaN;
+
+        private static Entity ApplylHopitalRule(Entity expr, Variable x, Entity dest)
+        {
+            if (expr is Divf(var num, var den))
+                if (num.Limit(x, dest).Evaled is var numLimit && den.Limit(x, dest).Evaled is var denLimit)
+                    if (numLimit == 0 && denLimit == 0 ||
+                            IsInfiniteNode(numLimit) && IsInfiniteNode(denLimit))
+                        if (num is not Number && den is not Number)
+                            if (num.ContainsNode(x) && den.ContainsNode(x))
+                            {
+                                var applied = num.Differentiate(x) / den.Differentiate(x);
+                                if (ComputeLimit(applied, x, dest) is { } resLim)
+                                    return resLim;
+                            }
+            return expr;
+        }
+
+        private static Entity ApplyTrivialTransformations(Entity expr, Variable x, Entity dest, Func<Entity, Entity, Entity> transformation)
+            => expr switch
+            {
+                Sumf(var a, var b)
+                    when ComputeLimit(a, x, dest) is { } aLim && ComputeLimit(b, x, dest) is { } bLim &&
+                        IsFiniteNode(aLim.Evaled) && IsFiniteNode(bLim.Evaled)
+                        => transformation(a, aLim) + transformation(b, bLim),
+                Minusf(var a, var b)
+                    when ComputeLimit(a, x, dest) is { } aLim && ComputeLimit(b, x, dest) is { } bLim &&
+                        IsFiniteNode(aLim.Evaled) && IsFiniteNode(bLim.Evaled)
+                        => transformation(a, aLim) - transformation(b, bLim),
+                Mulf(var a, var b)
+                    when ComputeLimit(a, x, dest) is { } aLim && ComputeLimit(b, x, dest) is { } bLim &&
+                        IsFiniteNode(aLim.Evaled) && IsFiniteNode(bLim.Evaled)
+                        => transformation(a, aLim) * transformation(b, bLim),
+                _ => expr
+            };
+
+        private static Entity TrivialTrigonometricReplacement(Entity expr, Variable x)
+            => expr switch
+            {
+                Secantf(var arg) when arg.ContainsNode(x) => 1 / MathS.Cos(arg),
+                Cosecantf(var arg) when arg.ContainsNode(x) => 1 / MathS.Sin(arg),
+                _ => expr
+            };
+
+    }
+}
